support macos in start-dev-env

+++ start-dev-env.sh
@@ -4,15 +4,29 @@ xhost +local:docker
 
 # Create SD card image if not exists:
 if [ ! -f sdcard.bin ]; then
-    echo "SD card image not found. Creating new image ..."
-    fallocate -l 50M sdcard.bin
-    sudo mkfs.fat ./sdcard.bin
+    if [[ "$(uname)" == "Darwin" ]]; then
+        echo "SD card image not found. Using blank image ..."
+        unzip sdcard.bin.zip
+    else
+        echo "SD card image not found. Creating new image ..."
+        fallocate -l 50M sdcard.bin
+        sudo mkfs.fat ./sdcard.bin
+    fi
 fi
 
 # Mount SD card image if not mounted:
-mount | grep -q sdcard.bin
-if [ "$?" -ne 0 ]; then
-    sudo mount -o loop ./sdcard.bin ./sdcard/
+
+if [[ "$(uname)" == "Darwin" ]]; then
+    mount | grep -q "$(pwd)"
+    if [ "$?" -ne 0 ]; then
+        MOUNTED_DISK=$(hdiutil attach -imagekey diskimage-class=CRawDiskImage -nomount ./sdcard.bin)
+        mount -t msdos $MOUNTED_DISK ./sdcard
+    fi
+else
+    mount | grep -q sdcard.bin
+    if [ "$?" -ne 0 ]; then
+        sudo mount -o loop ./sdcard.bin ./sdcard/
+    fi
 fi
 
 # Start container if not started:
